+++ x/token/keeper/keeper.go
@@ -4,6 +4,7 @@ import (
 	"github.com/line/lbm-sdk/codec"
 	sdk "github.com/line/lbm-sdk/types"
 	"github.com/line/lbm-sdk/x/token"
+	"github.com/line/lbm-sdk/x/token/class"
 )
 
 // Keeper defines the token module Keeper
@@ -31,4 +32,16 @@ func NewKeeper(
 		storeKey:      key,
 		cdc:           cdc,
 	}
+}
+
+func ValidateLegacyContract(k Keeper, ctx sdk.Context, contractID string) error {
+	if !k.classKeeper.HasID(ctx, contractID) {
+		return class.ErrContractNotExist.Wrap(contractID)
+	}
+
+	if _, err := k.GetClass(ctx, contractID); err != nil {
+		return err
+	}
+
+	return nil
 }